+++ src/python/impactx/dashboard/Run/__init__.py
@@ -1,4 +1,4 @@
-from .helper import SimulationHelper, SimulationProgress
+from .utils import SimulationHelper, SimulationProgress
 
 __all__ = [
     "SimulationHelper",
